+++ gitenberg/__init__.py
@@ -5,7 +5,10 @@
 
 from .book import Book
 from .clone import clone
+from .config import ConfigFile
+from .workflow import upload_all_books
 
-__title__ = 'gitenberg'
-__version__ = '0.0.9'
+__title__ = 'gitberg'
+__appname__ = 'gitberg'
+__version__ = '0.0.10'
 __copyright__ = 'Copyright 2012-2015 Seth Woodworth'