+++ HISTORY.rst
@@ -2,6 +2,11 @@
 
 History
 -------
+0.0.10 (2015-09-12)
+==================
+* Refactors and moves tests
+* Adds config file crossplatform using appdirs
+
 0.0.9 (2015-07-27)
 ==================
 * Adds clone command